react-ui-components: split the Html stories into sandbox and email suites
